Revert "[dhctl] feat(dhctl-for-commander): prepare dhctl for commander usage (#5893)"
This reverts commit 56e0b34f553594df73c1414ec8e579e42744bebc.
