+++ dhctl/pkg/infrastructure/cluster.go
@@ -17,7 +17,6 @@ package infrastructure
 import (
 	"github.com/deckhouse/deckhouse/dhctl/pkg/config"
 	"github.com/deckhouse/deckhouse/dhctl/pkg/kubernetes/actions/converge"
-	"github.com/deckhouse/deckhouse/dhctl/pkg/operations/phases"
 	"github.com/deckhouse/deckhouse/dhctl/pkg/state"
 )
 
@@ -37,24 +36,12 @@ type BaseInfraController interface {
 type ClusterInfra struct {
 	stateLoader StateLoader
 	cache       state.Cache
-
-	*phases.PhasedExecutionContext
 }
 
 func NewClusterInfra(terraState StateLoader, cache state.Cache) *ClusterInfra {
-	return NewClusterInfraWithOptions(terraState, cache, ClusterInfraOptions{})
-}
-
-type ClusterInfraOptions struct {
-	PhasedExecutionContext *phases.PhasedExecutionContext
-}
-
-func NewClusterInfraWithOptions(terraState StateLoader, cache state.Cache, opts ClusterInfraOptions) *ClusterInfra {
 	return &ClusterInfra{
 		stateLoader: terraState,
 		cache:       cache,
-
-		PhasedExecutionContext: opts.PhasedExecutionContext,
 	}
 }
 
@@ -67,14 +54,6 @@ func (r *ClusterInfra) DestroyCluster(autoApprove bool) error {
 	clusterState, nodesState, err := r.stateLoader.PopulateClusterState()
 	if err != nil {
 		return err
-	}
-
-	if r.PhasedExecutionContext != nil {
-		if shouldStop, err := r.PhasedExecutionContext.StartPhase(phases.AllNodesPhase, true); err != nil {
-			return err
-		} else if shouldStop {
-			return nil
-		}
 	}
 
 	for nodeGroupName, nodeGroupStates := range nodesState {
@@ -90,21 +69,5 @@ func (r *ClusterInfra) DestroyCluster(autoApprove bool) error {
 		}
 	}
 
-	if r.PhasedExecutionContext != nil {
-		if shouldStop, err := r.PhasedExecutionContext.SwitchPhase(phases.BaseInfraPhase, true, r.cache); err != nil {
-			return err
-		} else if shouldStop {
-			return nil
-		}
-	}
-
-	if err := NewBaseInfraController(metaConfig, r.cache).Destroy(clusterState, autoApprove); err != nil {
-		return err
-	}
-
-	if r.PhasedExecutionContext != nil {
-		return r.PhasedExecutionContext.CommitState(r.cache)
-	} else {
-		return nil
-	}
+	return NewBaseInfraController(metaConfig, r.cache).Destroy(clusterState, autoApprove)
 }
