--- dhctl/pkg/operations/phases/dhctl_state.go
@@ -1,30 +1,0 @@
-// Copyright 2023 Flant JSC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package phases
-
-import "github.com/deckhouse/deckhouse/dhctl/pkg/state"
-
-type DhctlState map[string][]byte
-
-func ExtractDhctlState(stateCache state.Cache) (res DhctlState, err error) {
-	err = stateCache.Iterate(func(k string, v []byte) error {
-		if res == nil {
-			res = make(map[string][]byte)
-		}
-		res[k] = v
-		return nil
-	})
-	return
-}
