+++ dhctl/pkg/state/cache/init.go
@@ -41,7 +41,7 @@ var (
 
 var globalCache state.Cache = &cache.DummyCache{}
 
-func choiceCache(identity string, opts CacheOptions) (state.Cache, error) {
+func choiceCache(identity string) (state.Cache, error) {
 	tmpDir := filepath.Join(app.CacheDir, stringsutil.Sha256Encode(identity))
 	log.DebugF("Cache dir %s\n", tmpDir)
 	if err := os.MkdirAll(tmpDir, 0o755); err != nil {
@@ -49,9 +49,6 @@ func choiceCache(identity string, opts CacheOptions) (state.Cache, error) {
 	}
 
 	if app.CacheKubeNamespace == "" {
-		if opts.InitialState != nil {
-			return cache.NewStateCacheWithInitialState(tmpDir, opts.InitialState)
-		}
 		return cache.NewStateCache(tmpDir)
 	}
 
@@ -90,24 +87,16 @@ func choiceCache(identity string, opts CacheOptions) (state.Cache, error) {
 	return k8sCache, nil
 }
 
-func initCache(identity string, opts CacheOptions) error {
+func initCache(identity string) error {
 	var err error
 	once.Do(func() {
-		globalCache, err = choiceCache(identity, opts)
+		globalCache, err = choiceCache(identity)
 	})
 	return err
 }
 
-type CacheOptions struct {
-	InitialState map[string][]byte
-}
-
 func Init(identity string) error {
-	return initCache(identity, CacheOptions{})
-}
-
-func InitWithOptions(identity string, opts CacheOptions) error {
-	return initCache(identity, opts)
+	return initCache(identity)
 }
 
 func Global() state.Cache {
