+++ dhctl/pkg/system/ssh/cmd/ssh-agent.go
@@ -110,9 +110,3 @@ func (a *SSHAgent) Start() error {
 	})
 	return nil
 }
-
-func (a *SSHAgent) Stop() {
-	if a.Executor != nil {
-		a.Executor.Stop()
-	}
-}
